출입문 관리 일괄등록 및 트리관련 오류 수정 및 기능보완 (#40)
* property 정리 및 세종청사 출입문 관련 property 항목 추가

* select query - door, building, floor 코드 추가

* Feat: 읽기모드(crudType) 분기 처리

* Feat: 읽기모드(crudType) 분기 처리

* Feat: 출입문관리 출입문, 층, 빌딩 조회 시 코드 번호 조회 추가

* Feat: 출입문관리 출입문, 층, 빌딩 조회 시 코드 번호 조회 추가

* 개인정보 영향평가 - 6번관련

* db설계변경 - 출입문관리 : door/detail.do에서 door_cd 추가

* Mod: 트리구조 읽기모드(R) 분기 처리/ 출입문,층 경로 읽기모드 분기

* Mod: 트리구조 읽기모드(R) 분기 처리/ 출입문,층 경로 읽기모드 분기

* Mod: 출입문 선택 팝업 트리구조 읽기모드(R) 분기 적용

* sql : building 조회시  -> area  컬럼 추가

* t_area table 미사용 관련 비즈니스 로직 update

* Fix: 출입문관리 - area 구역 관련 코드 삭제

* Mod: R모드 시 단말기 코드, 권한그룹 "선택" 버튼 안보이게 / 스케쥴, 알람그룹 "선택" --> "없음" 변경

* Fix: tree구조 area 제거, 샘플 엑셀 파일(excelupload_door_sample.xlsx)추가

* Fix: tree구조 area 제거, 샘플 엑셀 파일(excelupload_door_sample.xlsx)추가

* cubox 개발 db - profile(dev-cubox) 추가

* Fix: tree구조 area 제거, 샘플 엑셀 파일(excelupload_door_sample.xlsx)추가

* 출입문관리 - 출입문 상세 query 중복 제거 처리

* 출입문 엑셀업로드 관련 - 데이터 초기화 서비스 추가

* Mod: 출입문관리 엑셀 파일 업로드 로직 변경 중

* Mod: 출입문관리 엑셀 파일 업로드 로직 변경 중

* 출입문 관련 DB 변경 -  cd column 추가 대응

* Mod: 빌딩 선택 시 option 변경 로직 수정

* Mod: 출입문 일괄등록 엑셀 업로드 기능 변경 중

* Mod: 출입문관리 엑셀 파일 업로드 로직 변경 중

* Mod: 빌딩 선택 시 option 변경 로직 수정

* Mod: 출입문 일괄등록 엑셀 업로드 기능 변경 중

* Fix: 출입문관리 출입문 일괄등록 엑셀 업로드 기능 구현

* Mod: 출입문 일괄등록 엑셀 업로드 시 화면 잠금 반영

* fix - 출입문 그룹관리 update 오류 수정

* Mod: 출입문스케쥴 초 단위 중복 제거

* Fix: 출입문 알람그룹 등록 시 제목 오류, 출입문 목록 / 출입문 등록 팝업창 오류 수정

* Fix: 출입문 그룹관리 스케쥴명 링크 삭제

* fix - 출입문그룹관리

출입문 제외 오류 그룹명 등록시 오류

* Fix: 출입문 알람그룹 등록 시 제목 오류, 출입문 목록 / 출입문 등록 팝업창 오류 수정

* Fix: 출입문 그룹관리 스케쥴명 링크 삭제

* Fix: 출입문 일괄 등록 로직 수정

* Fix: 출입문 일괄 등록 로직 수정

* floor 정렬순서 지하 -> 옥상순으로 재정렬

* Fix: 출입문 일괄 등록 로직 수정

* Fix: blockUI.js 삭제

* Fix: blockUI.js 삭제

* 출입문 정렬순서 변경

* Mod: selectFloorList 층 정보 가져올 때 building_cd 추가

* Mod: selectFloorList 층 정보 가져올 때 building_cd 추가

* floor 정렬 순서 변경

* Mod: @transactional 태그 추가

* Mod: @transactional 태그 추가

* Mod: 출입문스케쥴 - 출입문그룹 선택 팝업 [권한그룹] 항목 삭제

* feat : 출입문 알람그룹 

출입분 알람그룹 상세 - 출입문 목록 door_nms 항목추가
출입문 알람그룹 상세 - 출입문 CRUD 기능 추가

* Feat: 출입문 알람 그룹 상세(수정), 추가 시 출입문 항목 추가 / 출입문 목록만 보는 팝업 제거

* Feat: 출입문 알람 그룹 상세(수정), 추가 시 출입문 항목 추가 / 출입문 목록만 보는 팝업 제거

* Feat: 출입문그룹관리 - 출입문 수 항목 추가

* Mod: 팝업 창 항목 선택, 제거 시 validation 추가

* Feat: 텍스트 input 특수문자 입력 제한

* Feat: 텍스트 input 특수문자 입력 제한

* Feat : 출입문 그룹 상세 - 출입문 수 조회 추가

* Fix: 출입문관리 - 단말기 선택 팝업 창 - 관리번호, 출입문 명 null 처리

* Feat: 텍스트 input 특수문자 입력 제한

* Fix: 출입문관리 - 단말기 선택 팝업 창 - 관리번호, 출입문 명 null 처리

* Fix: 출입문 일괄등록 양식 엑셀 수정

* Fix: 출입문 일괄등록 양식 엑셀 수정

* Mod: 출입문그룹관리 - 출입문 스케쥴 명 null 체크

* Mod: 요일 별 스케쥴 등록 validation 수정

* Mod: 출입문 관리 - 출입문 일괄등록 엑셀 업로드 validation 추가, ErrorCode 추가

* Mod: 출입문 관리 - 출입문 일괄등록 엑셀 업로드 validation 추가, ErrorCode 추가

* Mod: exception 처리 추가

* Mod: 오류코드 final 정의

* Mod: 출입문스케쥴 등록 시 항목 disabled 색상 통일

* fix : 출입문 알람그룹 등록시 출입문 데이터 추가

* Mod: 에러코드 추가 (ED03)

* Mod: 에러코드 추가 (ED03)

* Fix: 출입문알람그룹 출입문선택팝업창 닫을 시 전체선택 체크박스 선택 해제 반영

* Mod : 대쉬보드 구조 변경

* Feat: 대시보드 차트 구현 중

* Mod: 대시보드 차트 구현중

* Mod: 대시보드 차트 구현중

* Feat : 대쉬보드-  알람이력 및 출입이력 추가 /상단 바로가기 추가

* feat : 대쉬보드 차트 쿼리 생성

* Feat: 알람이력 차트 구현

* Feat: 알람이력 차트 구현

* Feat : 대쉬보드 알람이력현황 차트 쿼리 생성

* Mod: 대시보드 출입이력 현황 차트 서버 연동

* Mod: 출입이력현황 차트 서버 연동 수정

* Mod: 대시보드 알람이력 현황 수정

* 메인컨트롤러 변경 및 대쉬보드 구성 변경

* Mod: 대시보드 css 수정

* Mod: 출입문 엑셀 업로드 기능 서비스 분리

* no message

* Mod: 엑셀 업로드 - delete, insert 트랜잭션 합침

* Mod: 출입문엑셀업로드 - 단말기코드 반영

* Mod: 출입문 요일별 스케쥴 초 단위 비교 로직 추가

* Mod: 출입문 요일별 스케쥴 초 단위 비교 로직 추가

* Style: console.log 삭제

* Style: console.log 삭제

* Fix: 출입문 요일별 스케쥴 버그 수정

* Mod: 출입문스케쥴 출입문그룹 -> 출입문 변경

* Fix: 출입문 요일별 스케쥴 버그 수정

* Mod :   스케쥴 출입문 그룹관리 비즈니스 로직 수정(db 변경)

* Mod :   스케쥴 출입문 그룹관리 db 변경관련 비즈니스 로직 수정

* Mod: 1. 출입문스케쥴 - 출입문 컬럼 제거   2. 출입문관리 - 단말기선택창 사용여부 컬럼 추가   3. 출입문알람그룹 - 수정 후 취소 시 제목 오류 수정

* 스케쥴 관리 - 출입문 그룹 관련 로직 주석처리

* 메인화면 로고 background  color 변경 -> #FFFFFF

* dateitme(6) 대응 쿼리 수정

* (스케쥴 --> 스케쥴그룹) 관련된 변수 및 함수명 변경

* Mod: 출입문관리 (스케쥴 그룹, 스케쥴) 컬럼 추가

* Mod: 출입문관리 (스케쥴 그룹, 스케쥴) 컬럼 추가

* Mod: 알람그룹, 스케쥴 그룹, 스케쥴 값 체크 후 "없음" 표시

* Mod: 1. 출입문스케쥴 종료시간 최초 등록 시 초 단위 비교 로직 오류 수정  2. 출입문 엑셀 업로드 단말기 코드 validation 추가  3. 출입문등록 - 단말기 선택 시 이미 사용중인 단말기 연결 여부 확인하도록 로직 변경

* Style: schedule.detail 주석 정리

* Style: schedule.detail 주석 정리

* Mod: 출입문관리 - 사용중인 단말기 연결 여부 처리 수정 / 주석 제거

* Fix: 요일별스케쥴 버그 수정

* Fix: 층 입력 쿼리 수정 - buildingCd 입력

* Fix: 스케쥴 추가 시 사용여부 'useYn' parameter 추가

* Fix: 출입문관리 출입문 등록 시 출입문코드'doorCd' 추가

* Style: 주석 제거

* Mod: 출입문 스케쥴 수정 시 스케쥴그룹, 스케쥴 수정 제한 / 주석 정리

* Mod: 출입문알람그룹 - 추가된 DB 컬럼 alarm_use_typ(사용), door_alarm_typ(알람유형) 에 따른 CRUD 반영

* Mod: 스케쥴, 알람그룹 이름 유효성 체크 / 스케쥴 ui 변경

* Mod: 출입문 정보 수정 시 출입문코드(doorCd) 누락 반영 / 알람그룹 정보 저장, 수정 시 반영

* Fix: 출입문 등록 시 유효성 체크 후 disabled 해제 현상 수정

* Mod: 출입문 등록, 수정 시 doorCd 누락 반영 / 단말기 코드 선택 시 초기화 안됨 현상 수정

* Mod: 출입문관리 등록, 수정 시 빌딩코드, 층 코드 누락 반영

* Style: 주석 삭제

* Mod: 출입문 삭제 시 연결된 단말기, 권한그룹 삭제

* Style: 주석제거

* Mod: 빌딩코드, 층코드, 출입문코드 자릿수 제한

* Mod: 단말기 선택 창 - 단말기 변경 안하고 확인 눌렀을 때 버그 수정

* Mod: 출입문 스케쥴, 알람그룹 중복확인 버튼 css 수정

* Style: 코딩 컨벤션 수정

* Mod: 단말기 정보 초기화 이름 변경

* Mod: 출입문 알람그룹 조회, 등록, 수정 시 출입문 반영

* Style: console 추석 처리

* Style: console 주석 처리

* Style: console 주석 처리

* Mod: 출입문일괄등록 시 공백 체크 로직 추가

* Fix: 출입문 트리구조 첫번째 node 선택 안됨 현상 수정

* Mod: 출입문관리 대상 없는 추가 시 alert

* Mod: 출입문관리 권한그룹 불러오기 수정

---------

+++ src/main/java/aero/cubox/door/service/impl/DoorServiceImpl.java
@@ -56,6 +56,17 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
     @Transactional
     public String addDoor(Map<String, Object> commandMap) {
 
+        String doorCd = commandMap.get("doorCd").toString();
+
+        if (doorCd.length() < 6) {
+            String preNum = "";
+            int num = 6 - doorCd.length();
+            for (int j = 0; j < num; j++) {
+                preNum += "0";
+            }
+            commandMap.put("doorCd", preNum + doorCd);
+        }
+
         doorDAO.insertDoor(commandMap);
 
         String newDoorId = "";
@@ -63,37 +74,38 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
 
         HashMap paramMap = new HashMap();
 
-        if(!StringUtil.isEmpty(commandMap.get("doorId").toString() ) ){
+        if (!StringUtil.isEmpty(commandMap.get("doorId").toString())) {
 
             //도어그룹의 스케줄에 출입문 id Update
-            if( !StringUtil.isEmpty((String) commandMap.get("doorGroupId"))){
+            if (!StringUtil.isEmpty((String) commandMap.get("doorGroupId"))) {
 
-                paramMap.put("doorId", newDoorId );
+                paramMap.put("doorId", newDoorId);
                 paramMap.put("doorgrpId", commandMap.get("doorGroupId"));
 
                 doorGroupDAO.addDoorInDoorGroup(paramMap);
             }
 
             //단말기정보에 출입문 id Update
-            if( !StringUtil.isEmpty((String) commandMap.get("terminalIds"))){
+            if (!StringUtil.isEmpty((String) commandMap.get("terminalIds"))) {
 
-                paramMap.put("doorId", newDoorId );
+                paramMap.put("doorId", newDoorId);
                 paramMap.put("doorCd", commandMap.get("doorCd"));
                 paramMap.put("id", commandMap.get("terminalIds"));
+
                 doorDAO.updateDoorIdForTerminal(paramMap);
             }
 
             //출입권한-출입문 table에 door_id Insert
-            if( !StringUtil.isEmpty((String) commandMap.get("authGrIds"))){
+            if (!StringUtil.isEmpty((String) commandMap.get("authGrIds"))) {
 
                 String authGrIds = "";
                 authGrIds = commandMap.get("authGrIds").toString();
 
-                if( authGrIds.length() > 0 ){
+                if (authGrIds.length() > 0) {
                     String[] authGrIdArr = authGrIds.split("/");
                     for (int i = 0; i < authGrIdArr.length; i++) {
                         paramMap.put("authId", authGrIdArr[i]);
-                        paramMap.put("doorId", newDoorId );
+                        paramMap.put("doorId", newDoorId);
 
                         doorDAO.insertDoorIdForAuthDoor(paramMap);
                     }
@@ -113,34 +125,36 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
 
         HashMap paramMap = new HashMap();
 
-        
         //출입문그룹의 스케줄에 출입문 id - Insert or Update
-        if( !StringUtil.isEmpty((String) commandMap.get("doorGroupId"))){
+        if (!StringUtil.isEmpty((String) commandMap.get("doorGroupId"))) {
 
-            paramMap.put("doorId", commandMap.get("id") );
+            paramMap.put("doorId", commandMap.get("id"));
             paramMap.put("doorgrpId", commandMap.get("doorGroupId"));
+
             doorGroupDAO.deleteDoorInDoorGroup(paramMap);
             doorGroupDAO.addDoorInDoorGroup(paramMap);
         }
 
         //단말기정보에 출입문 id Update
-        if( !StringUtil.isEmpty((String) commandMap.get("terminalIds"))){
-            paramMap.put("doorId", commandMap.get("id") );
+        if (!StringUtil.isEmpty((String) commandMap.get("terminalIds"))) {
+
+            paramMap.put("doorId", commandMap.get("id"));
             paramMap.put("doorCd", commandMap.get("doorCd"));
             paramMap.put("id", commandMap.get("terminalIds"));
-            doorDAO.deleteDoorIdForTerminal(paramMap);
+
+            doorDAO.updateDoorIdForTerminalInit(paramMap);
             doorDAO.updateDoorIdForTerminal(paramMap);
         }
 
         //출입권한-출입문 table에 door_id Delete-Insert
-        if( !StringUtil.isEmpty((String) commandMap.get("authGrIds"))){
+        if (!StringUtil.isEmpty((String) commandMap.get("authGrIds"))) {
 
-            paramMap.put("doorId", commandMap.get("id") );
+            paramMap.put("doorId", commandMap.get("id"));
 
             String authGrIds = "";
             authGrIds = commandMap.get("authGrIds").toString();
 
-            if( authGrIds.length() > 0 ){
+            if( authGrIds.length() > 0 ) {
                 String[] authGrIdArr = authGrIds.split("/");
                 for (int i = 0; i < authGrIdArr.length; i++) {
                     paramMap.put("authId", authGrIdArr[i]);
@@ -164,23 +178,33 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
         HashMap paramMap = new HashMap();
 
         //단말기정보에 출입문 id - Update
-        if( !StringUtil.isEmpty((String) commandMap.get("terminalIds"))){
-            paramMap.put("doorId", doorId );
-            paramMap.put("id", commandMap.get("terminalIds"));
-            doorDAO.updateDoorIdForTerminal(paramMap);
+        if (!StringUtil.isEmpty((String) commandMap.get("terminalIds"))) {
+
+            paramMap.put("doorId", doorId);
+
+            doorDAO.updateDoorIdForTerminalInit(paramMap);
+//            doorDAO.updateDoorIdForTerminal(paramMap);
         }
 
         //출입권한-출입문 table에 door_id Delete-Insert
-        if( !StringUtil.isEmpty((String) commandMap.get("authGrIds"))){
+        if (!StringUtil.isEmpty((String) commandMap.get("authGrIds"))) {
 
-            paramMap.put("doorId", doorId );
-            paramMap.put("authId", commandMap.get("authGrIds"));
+            paramMap.put("doorId", doorId);
+            String authGrIds = "";
+            authGrIds = commandMap.get("authGrIds").toString();
 
-            doorDAO.deleteDoorIdForAuthDoor(paramMap);
+            if (authGrIds.length() > 0) {
+                String[] authGrIdArr = authGrIds.split("/");
+                for (int i = 0; i < authGrIdArr.length; i++) {
+                    paramMap.put("authId", authGrIdArr[i]);
+                    doorDAO.deleteDoorIdForAuthDoor(paramMap);
+                }
+            }
         }
 
         doorDAO.deleteDoor(commandMap);
     }
+
 
     @Override
     public void deleteDoorAll() {
@@ -219,6 +243,19 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
     @Override
     @Transactional
     public String addBuilding(HashMap paramMap) {
+
+        String buildingCd = paramMap.get("buildingCd").toString();
+
+        // buildingCd 2자리수 변형
+        if (buildingCd.length() < 2) {
+            String preNum = "";
+            int num = 2 - buildingCd.length();
+            for (int j = 0; j < num; j++) {
+                preNum += "0";
+            }
+            paramMap.put("buildingCd", preNum + buildingCd);
+        }
+
         doorDAO.insertBuilding(paramMap);
 
         String newBuildingId = "";
@@ -227,16 +264,16 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
         paramMap.put("buildingId", newBuildingId);
 
         //출입권한-빌딩 table에 buildingId Insert
-        if( !StringUtil.isEmpty((String) paramMap.get("authGrIds")) ){
+        if (!StringUtil.isEmpty((String) paramMap.get("authGrIds"))) {
 
             String authGrIds = "";
             authGrIds = paramMap.get("authGrIds").toString();
 
-            if( authGrIds.length() > 0 ){
+            if (authGrIds.length() > 0) {
                 String[] authGrIdArr = authGrIds.split("/");
                 for (int i = 0; i < authGrIdArr.length; i++) {
                     paramMap.put("authId", authGrIdArr[i]);
-                    paramMap.put("buildingId", newBuildingId );
+                    paramMap.put("buildingId", newBuildingId);
 
                     doorDAO.insertDoorIdForAuthDoor(paramMap);
                 }
@@ -244,6 +281,16 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
         }
         return newBuildingId;
     }
+
+    /**
+     * 빌딩코드 조회
+     * @param buildingId
+     * @return
+     */
+    public String getBuildingCd(String buildingId) {
+        return doorDAO.getBuildingCd(buildingId);
+    }
+
 
     /**
      * 출입문 수정
@@ -342,9 +389,21 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
     public Map getFloorDetail(Map<String, Object> paramMap) {
         return doorDAO.getFloorDetail(paramMap);
     }
+
     @Override
     @Transactional
     public String addFloor(HashMap paramMap) {
+        String floorCd = paramMap.get("floorCd").toString();
+
+        // floorCd 2자리수 변형
+        if (floorCd.length() < 2) {
+            String preNum = "";
+            int num = 2 - floorCd.length();
+            for (int j = 0; j < num; j++) {
+                preNum += "0";
+            }
+            paramMap.put("floorCd", preNum + floorCd);
+        }
         doorDAO.insertFloor(paramMap);
 
         String newFloorId = "";
@@ -366,6 +425,15 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
     @Override
     public void deleteFloorAll() {
         doorDAO.deleteFloorAll();
+    }
+
+    /**
+     * 층코드 조회
+     * @param floorId
+     * @return
+     */
+    public String getFloorCd(String floorId) {
+        return doorDAO.getFloorCd(floorId);
     }
 
     @Override
@@ -424,8 +492,8 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
             // 행이 없으면 패스
             if (row == null) continue;
 
-            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                               // 빌딩 명
-            String buildingCd = getValue(row.getCell(5)).replaceAll("\n", "<br>");                                               // 빌딩 코드
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>").trim();                                               // 빌딩 명
+            String buildingCd = getValue(row.getCell(5)).replaceAll("\n", "<br>").replaceAll("\\s", "");                                               // 빌딩 코드
 
             if (!buildingMap.containsValue(String.format("%02d", Integer.parseInt(buildingCd)))) { // buildingCd가 buildingMap에 없는 경우
                 buildingMap.put(buildingNm, String.format("%02d", Integer.parseInt(buildingCd)));
@@ -450,11 +518,11 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
             if (row == null) continue;
 
             HashMap<String, String> floorInfo = replaceFloorInfo
-                                (getValue(row.getCell(2)).replaceAll("\n", "<br>")
-                                , getValue(row.getCell(6)).replaceAll("\n", "<br>"));
+                                (getValue(row.getCell(2)).replaceAll("\n", "<br>").trim()
+                                , getValue(row.getCell(6)).replaceAll("\n", "<br>").replaceAll("\\s", ""));
 
-            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                            // 빌딩 명
-            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).replaceAll("\n", "<br>")));   // 빌딩 코드
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>").trim();                                            // 빌딩 명
+            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).replaceAll("\n", "<br>").replaceAll("\\s", "")));   // 빌딩 코드
             String buildingId = getBuildingId(buildingList, buildingNm, buildingCd);                                                                 // 빌딩 id
             String floorNm = floorInfo.get("floorNm");                                                                                               // 층 명
             String floorCd = floorInfo.get("floorCd");                                                                                               // 층 코드
@@ -483,16 +551,20 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
             Row row = sheet.getRow(i);
 
             if (row == null) continue;
+            HashMap<String, String> floorInfo = replaceFloorInfo(
+                      getValue(row.getCell(2)).replaceAll("\n", "<br>").trim()
+                    , getValue(row.getCell(6)).replaceAll("\n", "<br>").replaceAll("\\s", ""));
 
-            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>");                                              // 빌딩 명
-            String doorNm = getValue(row.getCell(3)).replaceAll("\n", "<br>");                                                  // 출입문 명
-            String terminalCd = getValue(row.getCell(4)).replaceAll("\n", "<br>");                                             // 단말기 코드
-            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).replaceAll("\n", "<br>")));     // 빌딩 코드 (2자리로 넣어야함)
-            String floorCd = getValue(row.getCell(6)).replaceAll("\n", "<br>");                                                 // 층 코드 (2자리로 넣어야함)
-            String doorCd = getValue(row.getCell(7)).replaceAll("\n", "<br>");                                                  // 출입문 코드
-            String authGrIds = getValue(row.getCell(8)).replaceAll("\n", "<br>");                                               // 빌딩 명
-            String buildingId = getBuildingId(buildingList, buildingNm, buildingCd);                                                                   // 빌딩 id
-            String floorId = getFloorId(floorList, floorCd, buildingCd);                                                                               // 층 id
+            String buildingNm = getValue(row.getCell(1)).replaceAll("\n", "<br>").trim();                                   // 빌딩 명
+            String doorNm = getValue(row.getCell(3)).replaceAll("\n", "<br>").trim();                                       // 출입문 명
+            String terminalCd = getValue(row.getCell(4)).replaceAll("\n", "<br>").replaceAll("\\s", "");    // 단말기 코드
+            String buildingCd = String.format("%02d", Integer.parseInt(getValue(row.getCell(5)).
+                                replaceAll("\n", "<br>").replaceAll("\\s", "")));                                  // 빌딩 코드 (2자리로 넣어야함)
+            String floorCd = floorInfo.get("floorCd");                                                                                             // 층 코드 (2자리로 넣어야함)
+            String doorCd = getValue(row.getCell(7)).replaceAll("\n", "<br>").replaceAll("\\s", "");        // 출입문 코드
+            String authGrIds = getValue(row.getCell(8)).replaceAll("\n", "<br>");                                           // 빌딩 명
+            String buildingId = getBuildingId(buildingList, buildingNm, buildingCd);                                                               // 빌딩 id
+            String floorId = getFloorId(floorList, floorCd, buildingCd);                                                                           // 층 id
 
             // doorCd 6자리수 변형
             if (doorCd.length() < 6) {
@@ -544,7 +616,7 @@ public class DoorServiceImpl extends EgovAbstractServiceImpl implements DoorServ
             floorInfo.put("floorNm", "PH층");
         }
         // floorCd 2자리수 변형
-        if (floorCd.length() == 1) {
+        if (floorCd.length() < 2) {
             floorInfo.put("floorCd", "0" + floorCd);
         }
         return floorInfo;
